load data directly from the json and move the images to the right place
Currently, this would not work well, because without an API key, the
JavaScript Geocoding API will block after more than 3 requests per second.
I will grab the data off the API for the dataset we are using, so we don't
have to include an API key for the demo.
It also works well live, but as mentioned, you need an API key.